+++ pandahub/api/internal/settings.py
@@ -1,10 +1,10 @@
 from dotenv import load_dotenv
 import os
 
-def settings_bool(var_name):
+def settings_bool(var_name, default=None):
     var = os.getenv(var_name)
     if var is None:
-        return None
+        return default
     if isinstance(var, str) and var.lower() == "true":
         return True
     elif isinstance(var, str) and var.lower() == "false":
@@ -15,6 +15,7 @@ def settings_bool(var_name):
 load_dotenv()
 
 MONGODB_URL = os.getenv("MONGODB_URL") or "mongodb://localhost:27017"
+MONGODB_URL_GLOBAL_DATABASE = os.getenv("MONGODB_URL_GLOBAL_DATABASE") or None
 
 EMAIL_VERIFICATION_REQUIRED = settings_bool("EMAIL_VERIFICATION_REQUIRED")
 
@@ -28,3 +29,7 @@ MAIL_SSL = os.getenv("MAIL_SSL") or False
 PASSWORD_RESET_URL = os.getenv("PASSWORD_RESET_URL") or ""
 EMAIL_VERIFY_URL = os.getenv("EMAIL_VERIFY_URL") or ""
 SECRET = os.getenv("SECRET")
+
+REGISTRATION_ENABLED = settings_bool("REGISTRATION_ENABLED", default=True)
+
+DATATYPES_MODULE = os.getenv("DATATYPES_MODULE") or "pandahub.lib.datatypes"