+++ pandahub/lib/database_toolbox.py
@@ -1,17 +1,13 @@
 # -*- coding: utf-8 -*-
-"""
-Created on Tue Feb  2 11:32:07 2021
 
-@author: julffers
-"""
-
+import numpy as np
 import pandas as pd
-import pandapower as pp
-import pandapipes as pps
+from pandahub.api.internal import settings
 import base64
 import hashlib
 import logging
 import json
+import importlib
 logger = logging.getLogger(__name__)
 
 
@@ -58,7 +54,7 @@ def make_task_hashable(task):
     return task
 
 
-def add_timestamp_info_to_document(document, timeseries):
+def add_timestamp_info_to_document(document, timeseries, ts_format):
     """
     Adds some meta information to documents that containg time series.
     The document will get the additional attributes 'first_timestamp',
@@ -80,12 +76,14 @@ def add_timestamp_info_to_document(document, timeseries):
         The updated timeseries metadata document.
 
     """
-    document["first_timestamp"] = timeseries.index[0]
-    document["last_timestamp"] = timeseries.index[-1]
+    if ts_format == "timestamp_value":
+        document["first_timestamp"] = timeseries.index[0]
+        document["last_timestamp"] = timeseries.index[-1]
     document["num_timestamps"] = len(timeseries.index)
     document["max_value"] = timeseries.max().item()
     document["min_value"] = timeseries.min().item()
     return document
+
 
 def convert_timeseries_to_subdocuments(timeseries):
     """
@@ -109,8 +107,38 @@ def convert_timeseries_to_subdocuments(timeseries):
                              "value": value})
     return subdocuments
 
-def create_timeseries_document(timeseries, data_type, element_type=None,
-                               netname=None, element_index=None, **kwargs):
+
+def compress_timeseries_data(timeseries_data, ts_format):
+    import blosc
+    if ts_format == "timestamp_value":
+        timeseries_data = np.array([timeseries_data.index.astype(int), 
+                                    timeseries_data.values])
+        return blosc.compress(timeseries_data.tobytes(),
+                              shuffle=blosc.SHUFFLE,
+                              cname="zlib")
+    elif ts_format == "array":
+        return blosc.compress(timeseries_data.astype(float).values.tobytes(),
+                              shuffle=blosc.SHUFFLE,
+                              cname="zlib")
+
+
+def decompress_timeseries_data(timeseries_data, ts_format):
+    import blosc
+    if ts_format == "timestamp_value":
+        data = np.frombuffer(blosc.decompress(timeseries_data), 
+                             dtype=np.float64).reshape((35040,2), 
+                                                       order="F")
+        return pd.Series(data[:,1], index=pd.to_datetime(data[:,0]))
+    elif ts_format == "array":
+        return np.frombuffer(blosc.decompress(timeseries_data), 
+                             dtype=np.float64)
+        
+
+def create_timeseries_document(timeseries, 
+                               data_type, 
+                               ts_format="timestamp_value",
+                               compress_ts_data=False,
+                               **kwargs):
     """
     Creates a document that contains timeseries metadata as well as the timeseries
     itself. Uses the function 'add_timestamp_info_to_document' to add information
@@ -149,21 +177,28 @@ def create_timeseries_document(timeseries, data_type, element_type=None,
         dict, representing a timeseries.
 
     """
-    document = {"data_type": data_type}
-    if element_type is not None:
-        document["element_type"] = element_type
-    if netname is not None:
-        document["netname"] = netname
-    if element_index is not None:
-        document["element_index"] = element_index
-    document = add_timestamp_info_to_document(document, timeseries)
+    document = {"data_type": data_type,
+                "ts_format": ts_format,
+                "compressed_ts_data": compress_ts_data}
+    document = add_timestamp_info_to_document(document, timeseries, ts_format)
     document = {**document, **kwargs}
     if not "_id" in document: # IDs set by users will not be overwritten
         document["_id"] = get_document_hash(document)
-    document["timeseries_data"] = convert_timeseries_to_subdocuments(timeseries)
+    if compress_ts_data:
+        document["timeseries_data"] = compress_timeseries_data(timeseries, 
+                                                               ts_format)
+    else:
+        if ts_format == "timestamp_value":
+            document["timeseries_data"] = convert_timeseries_to_subdocuments(timeseries)
+        elif ts_format == "array":
+            document["timeseries_data"] = list(timeseries.values)
+    
     return document
 
-def convert_dataframes_to_dicts(net, _id):
+def convert_dataframes_to_dicts(net, _id, datatypes=None):
+    if datatypes is None:
+        datatypes = getattr(importlib.import_module(settings.DATATYPES_MODULE), "datatypes")
+
     dataframes = {}
     other_parameters = {}
     types = {}
@@ -171,21 +206,41 @@ def convert_dataframes_to_dicts(net, _id):
         if key.startswith("_") or key.startswith("res"):
             continue
         if isinstance(data, pd.core.frame.DataFrame):
-            types[key] = {column: str(dtype) for column, dtype in net[key].dtypes.items()}
+
+            # ------------
+            # create type lookup
+
+            types[key] = dict()
+            default_dtypes = datatypes.get(key)
+            if default_dtypes is not None:
+                types[key].update({key: dtype.__name__ for key, dtype in default_dtypes.items()})
+            types[key].update(
+                {
+                    column: str(dtype) for column, dtype in net[key].dtypes.items()
+                    if column not in types[key]
+                }
+            )
             if data.empty:
                 continue
+
+            # ------------
             # convert pandapower objects in dataframes to dict
+
             df = net[key].copy(deep=True)
+
+            # ------------
+            # cast all columns with their default datatype
+
+            if default_dtypes is not None:
+                for column in df.columns:
+                    if column in default_dtypes:
+                        df[column] = df[column].astype(default_dtypes[column], errors="ignore")
+
             if "object" in df.columns:
-                df["object"] = df["object"].apply(
-                    lambda obj: {
-                        "_module": obj.__class__.__module__,
-                        "_class": obj.__class__.__name__,
-                        "_object": obj.to_json()
-                    }
-                )
+                df["object"] = df["object"].apply(object_to_json)
             df["index"] = df.index
             df["net_id"] = _id
+            load_geojsons(df)
             dataframes[key] = df.to_dict(orient="records")
         else:
             try:
@@ -195,3 +250,33 @@ def convert_dataframes_to_dicts(net, _id):
             else:
                 other_parameters[key] = data
     return dataframes, other_parameters, types
+
+def load_geojsons(df):
+    for column in df.columns:
+        if column == "geo" or column.endswith("_geo"):
+            df[column] = df[column].apply(lambda a: json.loads(a) if isinstance(a, str) else a)
+
+def convert_geojsons(df, geo_mode="string"):
+    if geo_mode == "dict":
+        return
+    for column in df.columns:
+        if column == "geo" or column.endswith("_geo"):
+            if geo_mode == "string":
+                df[column] = df[column].apply(lambda a: json.dumps(a) if isinstance(a, dict) else a)
+            elif geo_mode == "shapely":
+                from shapely.geometry import shape
+                df[column] = df[column].apply(lambda a: shape(a) if isinstance(a, dict) else a)
+            else:
+                raise NotImplementedError("Unknown geo_mode {}".format(geo_mode))
+
+def json_to_object(js):
+    _module = importlib.import_module(js["_module"])
+    _class = getattr(_module, js["_class"])
+    return _class.from_json(js["_object"])
+
+def object_to_json(obj):
+    return {
+        "_module": obj.__class__.__module__,
+        "_class": obj.__class__.__name__,
+        "_object": obj.to_json()
+    }