+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [0.2.3]- 2022-08-04
+
+    - ADDED version property in project data
+    - ADDED method to migrate projects to latest version
+    - ADDED option to disable registration
+    - ADDED option to use a separate mongodb instance as global database
+    - ADDED geo mode to handle geojson columns
+    - ADDED tutorials
+    - IMPROVED collections for element tables now start with 'net_'
+    - IMPROVED project IDs now can be any name
+    - IMPROVED compatibility with python < 3.9
+    - IMPROVED project settings API
+    - IMPROVED timeseries handling
+
 ## [0.2.2]- 2022-04-27
 
    - IMPROVED object support in network dataframes